client/requirements.txt
typing-extensions>=4.5.0
pytest>=7.3.1
websockets>=11.0.3
pyinstaller>=6.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
PyQt6>=6.5.0
darkdetect>=0.8.0